move PropertyCardView to Home, organize components folder
