+++ src/components/properties/NewPropertyCard.tsx
