+++ src/components/properties/PropertyCard.tsx
